using log(double(u)) in cu/propagate.cu aligns SC and AB positions, notes/issues/AB_SC_Position_Time_mismatch.rst but notes/issues/SC_Direction_mismatch.rst

+++ optixrap/cu/propagate.h
@@ -66,8 +66,10 @@ __device__ int propagate_to_boundary( Photon& p, State& s, curandState &rng)
     float u_boundary_burn = curand_uniform(&rng) ;
     float u_scattering = curand_uniform(&rng) ;
     float u_absorption = curand_uniform(&rng) ;
-    float scattering_distance = -s.material1.z*logf(u_scattering) ;     // .z:scattering_length
-    float absorption_distance = -s.material1.y*logf(u_absorption) ;     // .y:absorption_length 
+
+    float scattering_distance = -s.material1.z*log(double(u_scattering)) ;   // .z:scattering_length
+    float absorption_distance = -s.material1.y*log(double(u_absorption)) ;   // .y:absorption_length 
+    //  see notes/issues/AB_SC_Position_Time_mismatch.rst
 #else
     float scattering_distance = -s.material1.z*logf(curand_uniform(&rng));   // .z:scattering_length
     float absorption_distance = -s.material1.y*logf(curand_uniform(&rng));   // .y:absorption_length
